requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-standardized-errors==0.12.5
exceptiongroup==1.1.1
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
Markdown==3.4.3
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.3.2
pytest-django==4.5.2
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tomli==2.0.1
uritemplate==4.1.1
yarl==1.9.2
